Cambios en la liquidacion

+++ cromo/ApiPrecioDetalle.cs
@@ -9,7 +9,7 @@ namespace cromo
     class ApiPrecioDetalle
     {
         public string codigoPrecioDetallePk { get; set; }
-        public double minimo { get; set; }
+        public int minimo { get; set; }
         public double vrPeso { get; set; }
         public double vrUnidad { get; set; }
         public double vrPesoTope { get; set; }
